+++ src/components/tournament-overview-table/tournament-overview-table.component.js
@@ -7,5 +7,10 @@ module.exports = {
             overview: _.filter(input.overview, player => input.players[player.player_id].name !== ""),
             hasStatistics: !_.isEmpty(input.statistics)
         }
+    },
+    onMount() {
+        $(function() {
+            $('[data-toggle="tooltip"]').tooltip();
+          });
     }
 }